+++ customer-escalation-management/src/main/java/com/chickenfarms/escalationmanagement/repository/ProblemRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProblemRepository extends JpaRepository<Problem, Integer> {
+public interface ProblemRepository extends JpaRepository<Problem, Long> {
   Optional<Problem> findProblemByName(String name);
 }
