+++ customer-escalation-management/src/main/java/com/chickenfarms/escalationmanagement/repository/RootCauseRepository.java
@@ -3,5 +3,5 @@ package com.chickenfarms.escalationmanagement.repository;
 import com.chickenfarms.escalationmanagement.model.entity.RootCause;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RootCauseRepository extends JpaRepository<RootCause, Integer> {
+public interface RootCauseRepository extends JpaRepository<RootCause, Long> {
 }
